+++ pkg/storage/cache/doc.go
@@ -1,2 +1,2 @@
-// Package cache allows you to store values in the temporary storage.
+// Package cache allows you to store values in the inmemory storage.
 package cache