+++ pkg/storage/cache/types.go
@@ -1,5 +1,12 @@
 package cache
 
+type ILRUCache interface {
+	ICache
+
+	GetIndex() uint64
+	GetKey(i uint64) ([]byte, bool)
+}
+
 type ICache interface {
 	ICacheSetter
 	ICacheGetter